feat(markdown): optimise the markdown style - markdown style applied with module css; - refactor copy button with mui
